requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
beautifulsoup4==4.12.2
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2.1
django-allauth==0.54.0
django-bootstrap4==23.1
django-ckeditor==6.5.1
django-filter==23.2
django-js-asset==2.0.0
idna==3.4
kombu==5.2.4
oauthlib==3.2.2
prompt-toolkit==3.0.38
pycparser==2.21
PyJWT==2.7.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
